requirements.txt
absl-py==0.15.0
aiohttp==3.7.4.post0
alembic==1.7.5
astunparse==1.6.3
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==21.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
cachetools==5.0.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.1
colorama==0.4.4
cowin-vaccine-api==0.0.3
cryptography==3.4.7
cycler==0.11.0
DateTime==4.3
Deprecated==1.2.12
discord.py==1.7.2
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.3
entrypoints==0.3
fake-useragent==0.1.11
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
flatbuffers==2.0
gast==0.4.0
gcloud==0.17.0
google-auth==2.4.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.53.0
greenlet==1.1.0
grpcio==1.43.0
gunicorn==20.1.0
h5py==3.6.0
httplib2==0.19.1
idna==2.10
imbalanced-learn==0.8.0
imblearn==0.0
importlib-metadata==4.10.1
iniconfig==1.1.1
ipython-genutils==0.2.0
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-core==4.7.1
jwcrypto==1.0
jws==0.1.3
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
lxml==4.6.3
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.4.3
mediapipe==0.8.9
multidict==5.1.0
multitasking==0.0.9
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.6.2
numpy==1.21.0
oauth2client==3.0.0
oauthlib==3.1.1
opencv-contrib-python==4.5.4.58
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.5
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.1
pluggy==0.13.1
prompt-toolkit==3.0.20
protobuf==3.17.3
psutil==5.9.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.4.3
Pygments==2.10.0
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
Pyrebase==3.0.27
pyrsistent==0.18.0
pyTelegramBotAPI==3.7.9
pytest==6.2.3
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.17.1
python-jwt==2.0.1
pytz==2021.1
pywin32==301
pyzmq==22.1.0
razorpay==0.2.0
regex==2021.7.6
requests==2.27.1
requests-oauthlib==1.3.0
requests-toolbelt==0.7.0
rsa==4.7.2
scikit-learn==1.0
scipy==1.7.1
six==1.15.0
soupsieve==2.2.1
SQLAlchemy==1.4.17
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
tflearn==0.5.0
threadpoolctl==2.2.0
toml==0.10.2
tornado==6.1
tqdm==4.61.2
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.5
uuid==1.30
visitor==0.1.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
wrapt==1.12.1
WTForms==2.3.3
yarl==1.6.3
yfinance==0.1.59
zipp==3.7.0
zope.interface==5.4.0
